♻️ re-organize functions in 3 differents components for more legibility
